build: make FOTA the default firmware layout
