+++ Kconfig.sysbuild
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
 config NRF_DEFAULT_IPC_RADIO
 	default y
