+++ boards/openearable_v2_nrf5340_cpuapp.overlay
@@ -1,0 +1,35 @@
+/ {
+	chosen {
+		nordic,pm-ext-flash = &mx25r64;
+	};
+};
+
+&qspi {
+	status = "disabled";
+};
+
+&spi4 {
+	cs-gpios = <&gpio0 11 GPIO_ACTIVE_LOW>,<&gpio1 1 GPIO_ACTIVE_LOW>;
+	status = "okay";
+	mx25r64: mx25r6435f@1 {
+		compatible = "jedec,spi-nor";
+		reg = <1>;
+		spi-max-frequency = <32000000>;
+
+		jedec-id = [c2 28 17];
+		sfdp-bfp = [
+			e5 20 f1 ff  ff ff ff 03  44 eb 08 6b  08 3b 04 bb
+			ee ff ff ff  ff ff 00 ff  ff ff 00 ff  0c 20 0f 52
+			10 d8 00 ff  23 72 f5 00  82 ed 04 cc  44 83 68 44
+			30 b0 30 b0  f7 c4 d5 5c  00 be 29 ff  f0 d0 ff ff
+		];
+		size = <67108864>;
+		has-dpd;
+		t-enter-dpd = <10000>;
+		t-exit-dpd = <5000>;
+		dpd-wakeup-sequence = <30000 20 45000>;
+		mxicy,mx25r-power-mode = "high-performance";
+
+		zephyr,pm-device-runtime-auto;
+	};
+};
