--- cdc_acm_uart0_console.overlay
@@ -1,19 +1,0 @@
-// To get started, press Ctrl+Space (or Option+Esc) to bring up the completion menu and view the available nodes.
-
-// You can also use the buttons in the sidebar to perform actions on nodes.
-// Actions currently available include:
-
-// * Enabling / disabling the node
-// * Adding the bus to a bus
-// * Removing the node
-// * Connecting ADC channels
-
-// For more help, browse the DeviceTree documentation at https://docs.zephyrproject.org/latest/guides/dts/index.html
-// You can also visit the nRF DeviceTree extension documentation at https://docs.nordicsemi.com/bundle/nrf-connect-vscode/page/guides/ncs_configure_app.html#devicetree-support-in-the-extension
-
-/ {
-	chosen {
-		zephyr,console = &cdc_acm_uart0;
-		zephyr,shell-uart = &cdc_acm_uart0;
-    };
-};
