--- prj_fota.conf
@@ -1,263 +1,0 @@
-#
-# Copyright (c) 2022 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-# CONFIG_STATIC_INIT_GNU=y
-
-# nRF5340 Audio
-CONFIG_NRF5340_AUDIO=y
-CONFIG_SAMPLE_RATE_CONVERTER=y
-CONFIG_SAMPLE_RATE_CONVERTER_FILTER_SIMPLE=y
-
-# General
-CONFIG_DEBUG=y
-CONFIG_DEBUG_INFO=y
-CONFIG_ASSERT=y
-CONFIG_STACK_USAGE=y
-CONFIG_THREAD_RUNTIME_STATS=y
-CONFIG_STACK_SENTINEL=y
-CONFIG_INIT_STACKS=y
-
-# Uart driver
-CONFIG_SERIAL=y
-
-# Logging
-CONFIG_LOG=y
-CONFIG_NEWLIB_LIBC_FLOAT_PRINTF=y
-CONFIG_LOG_TAG_MAX_LEN=2
-CONFIG_LOG_TAG_DEFAULT="--"
-CONFIG_LOG_BACKEND_UART=y
-CONFIG_LOG_BUFFER_SIZE=4096
-
-# Use this for debugging thread usage
-#CONFIG_LOG_THREAD_ID_PREFIX=y
-
-# Console related defines
-CONFIG_CONSOLE=y
-CONFIG_RTT_CONSOLE=y
-CONFIG_UART_CONSOLE=y
-CONFIG_STDOUT_CONSOLE=y
-
-# Shell related defines
-CONFIG_SHELL=y
-CONFIG_KERNEL_SHELL=y
-CONFIG_USE_SEGGER_RTT=y
-## Disable logs on RTT
-CONFIG_SHELL_RTT_INIT_LOG_LEVEL_NONE=y
-CONFIG_SHELL_BACKEND_RTT=y
-CONFIG_SHELL_BACKEND_SERIAL=n
-CONFIG_SHELL_VT100_COMMANDS=y
-CONFIG_SHELL_VT100_COLORS=y
-CONFIG_SHELL_STACK_SIZE=4096
-CONFIG_SHELL_CMD_BUFF_SIZE=128
-## Reduce shell memory usage
-CONFIG_SHELL_WILDCARD=n
-CONFIG_SHELL_HELP_ON_WRONG_ARGUMENT_COUNT=n
-CONFIG_SHELL_STATS=n
-CONFIG_SHELL_CMDS=n
-CONFIG_SHELL_HISTORY=y
-
-# Turn off default shell commands
-CONFIG_I2C_SHELL=n
-CONFIG_HWINFO_SHELL=n
-CONFIG_CLOCK_CONTROL_NRF_SHELL=n
-CONFIG_FLASH_SHELL=n
-CONFIG_DEVICE_SHELL=n
-
-# Suppress LOG_ERR messages from sd_check_card_type. Because SPI_SDHC has no card presence method,
-# assume card is in slot. Thus error message is always shown if card is not inserted
-CONFIG_SD_LOG_LEVEL_OFF=y
-
-# Suppress LOG_INF messages from hci_core
-CONFIG_BT_HCI_CORE_LOG_LEVEL_WRN=y
-
-# DFU
-CONFIG_AUDIO_BT_MGMT_DFU=y
-CONFIG_MCUMGR_TRANSPORT_BT_PERM_RW=y
-CONFIG_BT_L2CAP_TX_MTU=498
-CONFIG_BT_BUF_ACL_TX_SIZE=251
-
-# External Flash
-CONFIG_FLASH=y
-CONFIG_FLASH_MAP=y
-CONFIG_SPI_NOR=y
-CONFIG_SPI_NOR_SFDP_DEVICETREE=y
-
-# Michael: missing somehow
-CONFIG_NRF5340_AUDIO_CS47L63_DRIVER=y
-CONFIG_NRF5340_AUDIO_POWER_MEASUREMENT=n
-CONFIG_NRF5340_AUDIO_SD_CARD_MODULE=y
-
-#Custom
-CONFIG_NRFX_PWM1=y
-
-CONFIG_I2C=y
-CONFIG_I2C_NRFX=y
-
-CONFIG_POWEROFF=y
-
-#CONFIG_PM=y
-CONFIG_PM_DEVICE=y
-CONFIG_PM_DEVICE_RUNTIME=y
-
-CONFIG_BT_DEVICE_NAME="OpenEarable"
-
-#CONFIG_USB_CDC_ACM_LOG_LEVEL_DBG=y
-
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_OBSERVER=y
-
-CONFIG_SFLOAT=y
-
-CONFIG_FPU=y
-
-CONFIG_NEWLIB_LIBC=y
-CONFIG_NEWLIB_LIBC_FLOAT_PRINTF=y
-CONFIG_NEWLIB_LIBC_FLOAT_SCANF=y
-
-CONFIG_CPP=y
-CONFIG_GLIBCXX_LIBCPP=y
-CONFIG_STD_CPP17=y
-
-#CONFIG_USB=y
-CONFIG_USB_CDC_ACM=y
-
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_USB_DEVICE_PRODUCT="OpenEarable v2"
-CONFIG_USB_DEVICE_PID=0x0004
-CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
-
-CONFIG_UART_LINE_CTRL=y
-
-CONFIG_BT_USER_DATA_LEN_UPDATE=y
-CONFIG_BT_USER_PHY_UPDATE=y
-
-CONFIG_BT_CTLR_DATA_LENGTH_MAX=251
-CONFIG_BT_BUF_ACL_RX_SIZE=502
-CONFIG_BT_BUF_ACL_TX_SIZE=502
-CONFIG_BT_L2CAP_TX_MTU=498
-CONFIG_BT_CTLR_PHY_2M=y
-CONFIG_BT_CTLR_PHY_CODED=n
-
-CONFIG_BT_ISO_TX_BUF_COUNT=4
-CONFIG_BT_ISO_RX_BUF_COUNT=8
-
-CONFIG_BT_CTLR_RX_BUFFERS=8
-#CONFIG_BT_CTLR_TX_BUFFERS=8
-
-CONFIG_BT_CTLR_ISO_TX_BUFFERS=16
-CONFIG_BT_CTLR_ISO_RX_BUFFERS=16
-
-#CONFIG_BT_ISO_RX_MTU=498
-#CONFIG_BT_ISO_TX_MTU=498
-
-CONFIG_BT_ATT_PREPARE_COUNT=8
-CONFIG_BT_L2CAP_DYNAMIC_CHANNEL=y
-
-#CONFIG_BT_AUDIO_PRESENTATION_DELAY_US=40000
-#CONFIG_BT_AUDIO_MAX_TRANSPORT_LATENCY_MS=20
-
-CONFIG_BT_AUDIO_RETRANSMITS=2
-
-CONFIG_BLE_ACL_CONN_INTERVAL=8
-
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_OBSERVER=y
-
-CONFIG_SFLOAT=y
-
-#CONFIG_RESET_ON_FATAL_ERROR=y
-
-CONFIG_STREAM_BIDIRECTIONAL=y
-
-# CONFIG_NRFX_PDM=y
-
-CONFIG_AUDIO_MIC_I2S=y
-
-# CONFIG_LOG_DEFAULT_LEVEL=4
-
-CONFIG_SPI_NRFX_RAM_BUFFER_SIZE=4096
-
-CONFIG_SPI=y
-CONFIG_SPI_NRFX=y
-CONFIG_DISK_ACCESS=y
-CONFIG_SDHC=y
-
-CONFIG_LOG=y
-CONFIG_FILE_SYSTEM=y
-CONFIG_FAT_FILESYSTEM_ELM=y
-
-# Enable SDHC interface
-CONFIG_DISK_DRIVERS=y
-CONFIG_DISK_DRIVER_SDMMC=y
-
-CONFIG_FS_FATFS_MOUNT_MKFS=n # Disable automatic formatting
-CONFIG_FS_FATFS_EXFAT=y      # Enable exFAT support
-
-CONFIG_FDSP=y
-
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=4096
-
-#CONFIG_USB_MASS_STORAGE=y
-#CONFIG_MASS_STORAGE_STACK_SIZE=1024
-
-##################################################
-
-# Step 1.1 - Enable SPI driver for the application
-#CONFIG_GPIO=y
-#CONFIG_SPI=y
-#CONFIG_SPI_NOR=y
-#CONFIG_SPI_NOR_SFDP_DEVICETREE=y
-CONFIG_SPI_NOR_FLASH_LAYOUT_PAGE_SIZE=4096
-# QSPI drivers are enabled by defualt for some chips.
-# Disable it explicitly to be sure QSPI is disabled.
-CONFIG_NORDIC_QSPI_NOR=n
-
-# STEP 5.3 - Enable USB subsystem
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_UART_LINE_CTRL=y
-CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
-
-# STEP 5.2 - Enable mcumgr DFU in application
-# Enable MCUMGR 
-CONFIG_MCUMGR=y
-
-# Enable MCUMGR management for both OS and Images
-CONFIG_MCUMGR_GRP_OS=y
-CONFIG_MCUMGR_GRP_IMG=y
-
-# Configure MCUMGR transport to UART
-CONFIG_MCUMGR_TRANSPORT_UART=y
-
-# Dependencies
-# Configure dependencies for CONFIG_MCUMGR  
-CONFIG_NET_BUF=y
-CONFIG_ZCBOR=y
-CONFIG_CRC=y
-
-# Configure dependencies for CONFIG_MCUMGR_GRP_IMG  
-CONFIG_FLASH=y
-CONFIG_IMG_MANAGER=y
-
-# Configure dependencies for CONFIG_IMG_MANAGER  
-CONFIG_STREAM_FLASH=y
-CONFIG_FLASH_MAP=y
-
-# Configure dependencies for CONFIG_MCUMGR_TRANSPORT_UART 
-CONFIG_BASE64=y
-
-# MCUboot hooks
-CONFIG_MCUMGR_MGMT_NOTIFICATION_HOOKS=y
-
-CONFIG_MCUMGR_SMP_COMMAND_STATUS_HOOKS=y
-#CONFIG_MCUMGR_GRP_FS_FILE_ACCESS_HOOK=y
-
-CONFIG_MCUMGR_GRP_IMG_UPLOAD_CHECK_HOOK=y
-CONFIG_MCUMGR_GRP_IMG_STATUS_HOOKS=y
-CONFIG_MCUMGR_GRP_OS_RESET_HOOK=y
-
-#CONFIG_MCUMGR_GRP_SETTINGS_ACCESS_HOOK=y
-
-CONFIG_SETUP_FUEL_GAUGE=y
