+++ prj_release.conf
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2022 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
 # nRF5340 Audio
 CONFIG_NRF5340_AUDIO=y
