--- sample.yaml
@@ -1,25 +1,0 @@
-sample:
-  name: nRF5340 Audio application
-  description: LE Audio and Auracast implementation example
-common:
-  integration_platforms:
-    - openearable_v2/nrf5340/cpuapp
-  platform_allow: openearable_v2/nrf5340/cpuapp
-  sysbuild: true
-  build_only: true
-  tags: ci_build sysbuild
-tests:
-  applications.nrf5340_audio.default:
-    extra_args: []
-  applications.nrf5340_audio.headset_unicast:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=1
-  applications.nrf5340_audio.gateway_unicast:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=2
-  applications.nrf5340_audio.headset_broadcast:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=1 CONFIG_TRANSPORT_BIS=y
-  applications.nrf5340_audio.gateway_broadcast:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=2 CONFIG_TRANSPORT_BIS=y
-  applications.nrf5340_audio.headset_unicast_sd_card:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=1 CONFIG_SD_CARD_PLAYBACK=y
-  applications.nrf5340_audio.headset_dfu:
-    extra_args: FILE_SUFFIX=release CONFIG_AUDIO_DEV=1 FILE_SUFFIX=fota
