+++ sysbuild.conf
@@ -1,9 +1,3 @@
-#
-# Copyright (c) 2024 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
 # Enable bootloaders for both cores
 SB_CONFIG_BOOTLOADER_MCUBOOT=y
 SB_CONFIG_SECURE_BOOT_NETCORE=y
