+++ sysbuild/mcuboot/boards/openearable_v2_nrf5340_cpuapp.overlay
@@ -1,0 +1,7 @@
+#include "../../../boards/openearable_v2_nrf5340_cpuapp.overlay"
+
+/ {
+	chosen {
+		nordic,pm-ext-flash = &mx25r64;
+	};
+};
