--- sysbuild/mcuboot/boards/openearable_v2_nrf5340_cpuapp_fota.overlay
@@ -1,7 +1,0 @@
-#include "../../../boards/openearable_v2_nrf5340_cpuapp_fota.overlay"
-
-/ {
-	chosen {
-		nordic,pm-ext-flash = &mx25r64;
-	};
-};
