--- unicast_server/CMakeLists.txt
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2023 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-target_sources(app PRIVATE
-	       ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp)
