--- unicast_server/overlay-unicast_server.conf
@@ -1,49 +1,0 @@
-#
-# Copyright (c) 2025 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-CONFIG_AUDIO_DEV=1
-
-CONFIG_BT_GATT_AUTO_DISCOVER_CCC=y
-CONFIG_BT_GATT_AUTO_RESUBSCRIBE=n
-CONFIG_BT_GATT_AUTO_SEC_REQ=n
-CONFIG_BT_GATT_AUTO_UPDATE_MTU=y
-CONFIG_BT_GATT_CACHING=n
-CONFIG_BT_GATT_CLIENT=y
-CONFIG_BT_GATT_DYNAMIC_DB=y
-
-CONFIG_BT_MAX_CONN=4
-CONFIG_BT_MAX_PAIRED=4
-CONFIG_BT_EXT_ADV=y
-
-CONFIG_BT_AUDIO=y
-
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_ISO_PERIPHERAL=y
-CONFIG_BT_BAP_UNICAST_SERVER=y
-CONFIG_BT_CAP_ACCEPTOR=y
-CONFIG_BT_CAP_ACCEPTOR_SET_MEMBER=y
-CONFIG_BT_PAC_SNK=y
-CONFIG_BT_PAC_SNK_NOTIFIABLE=y
-CONFIG_BT_PAC_SRC=y
-CONFIG_BT_PAC_SRC_NOTIFIABLE=y
-CONFIG_BT_CSIP_SET_MEMBER=y
-CONFIG_BT_DEVICE_APPEARANCE=2369
-
-CONFIG_BT_ISO_MAX_CHAN=2
-CONFIG_BT_ASCS=y
-CONFIG_BT_ASCS_MAX_ASE_SNK_COUNT=1
-CONFIG_BT_ASCS_MAX_ASE_SRC_COUNT=1
-CONFIG_BT_VCP_VOL_REND=y
-CONFIG_BT_MCC=y
-CONFIG_BT_MCC_READ_MEDIA_STATE=y
-CONFIG_BT_MCC_SET_MEDIA_CONTROL_POINT=y
-CONFIG_BT_GAP_PERIPHERAL_PREF_PARAMS=n
-CONFIG_BT_AUDIO_CODEC_CFG_MAX_METADATA_SIZE=25
-
-CONFIG_LC3_ENC_CHAN_MAX=1
-CONFIG_LC3_DEC_CHAN_MAX=1
-CONFIG_MBEDTLS_ENABLE_HEAP=y
-CONFIG_MBEDTLS_HEAP_SIZE=2048
